requirements.txt
python-gitlab>=2.11.0
